Moved basic I/O manipulators from base/Raw to base/IoManip
