+++ src/acl/ConnMark.cc
@@ -11,7 +11,7 @@
 #include "squid.h"
 #include "acl/ConnMark.h"
 #include "acl/FilledChecklist.h"
-#include "base/Raw.h"
+#include "base/IoManip.h"
 #include "client_side.h"
 #include "debug/Stream.h"
 #include "http/Stream.h"
