+++ src/base/Makefile.am
@@ -35,6 +35,8 @@ libbase_la_SOURCES = \
 	Here.h \
 	InstanceId.cc \
 	InstanceId.h \
+	IoManip.cc \
+	IoManip.h \
 	JobWait.cc \
 	JobWait.h \
 	Lock.h \
