+++ src/base/Raw.cc
@@ -6,29 +6,12 @@
  * Please see the COPYING and CONTRIBUTORS files for details.
  */
 
-/* DEBUG: section 00    Debug Routines */
-
 #include "squid.h"
+#include "base/IoManip.h"
 #include "base/Raw.h"
 #include "debug/Stream.h"
 
-#include <algorithm>
-
-void
-PrintHex(std::ostream &os, const char *data, const size_t n)
-{
-    if (!n)
-        return;
-    assert(data);
-
-    const auto savedFill = os.fill('0');
-    const auto savedFlags = os.flags(); // std::ios_base::fmtflags
-    os << std::hex;
-    std::for_each(data, data + n,
-    [&os](const char &c) { os << std::setw(2) << static_cast<uint8_t>(c); });
-    os.flags(savedFlags);
-    os.fill(savedFill);
-}
+#include <iostream>
 
 std::ostream &
 Raw::print(std::ostream &os) const
