+++ src/security/ErrorDetail.cc
@@ -7,7 +7,7 @@
  */
 
 #include "squid.h"
-#include "base/Raw.h"
+#include "base/IoManip.h"
 #include "error/SysErrorDetail.h"
 #include "html_quote.h"
 #include "sbuf/SBuf.h"
