+++ src/security/Io.cc
@@ -9,7 +9,7 @@
 /* DEBUG: section 83    TLS I/O */
 
 #include "squid.h"
-#include "base/Raw.h"
+#include "base/IoManip.h"
 #include "fde.h"
 #include "security/Io.h"
