+++ src/security/PeerConnector.cc
@@ -10,7 +10,7 @@
 
 #include "squid.h"
 #include "acl/FilledChecklist.h"
-#include "base/Raw.h"
+#include "base/IoManip.h"
 #include "comm/Loops.h"
 #include "comm/Read.h"
 #include "Downloader.h"
